templates: move DotNetNew machinery into Stride.Assets
WPF dialog injection becomes an IDotNetNewParameterPrompt constructor
parameter (WpfDotNetNewParameterPrompt adapter stays in Presentation).
SdtplMetadata also moves here and is linked from the CLI tool.
