+++ samples/Tests/Runner/ScreenshotRunner.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using System.Text.RegularExpressions;
 using Stride.Assets.Presentation;
 using Stride.Assets.Presentation.Templates;
+using Stride.Assets.Templates;
 using Stride.Core.Assets;
 using Stride.Core.Assets.Templates;
 using Stride.Core.Diagnostics;
