+++ sources/editor/Stride.Assets.Presentation/Templates/DotNetNewTemplateParametersWindow.xaml.cs
@@ -9,6 +9,7 @@ using System.Windows.Controls;
 using System.Windows.Media;
 using TextBox = System.Windows.Controls.TextBox;
 using Microsoft.TemplateEngine.Abstractions;
+using Stride.Assets.Templates;
 using Stride.Core.Presentation.Controls;
 using Stride.Data;
 using SDDialogResult = Stride.Core.Presentation.Services.DialogResult;
