+++ sources/editor/Stride.Assets.Presentation/Templates/StrideTemplates.cs
@@ -1,5 +1,6 @@
 // Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
+using Stride.Assets.Templates;
 using Stride.Core.Assets.Quantum.Templates;
 using Stride.Core.Assets.Templates;
 
@@ -15,7 +16,7 @@ namespace Stride.Assets.Presentation.Templates
             // asset graph before save. CLI / headless flows (Samples Tests, sample regen)
             // register session generators directly via TemplateManager.Register and skip the
             // wrap.
-            QuantumTemplateRegistration.Register(new DotNetNewTemplateGenerator());
+            QuantumTemplateRegistration.Register(new DotNetNewTemplateGenerator(new WpfDotNetNewParameterPrompt()));
             QuantumTemplateRegistration.Register(ProjectLibraryTemplateGenerator.Default);
             TemplateManager.Register(UpdatePlatformsTemplateGenerator.Default);
             TemplateManager.Register(AssetFactoryTemplateGenerator.Default);
