+++ sources/engine/Stride.Assets/Stride.Assets.csproj
@@ -23,6 +23,9 @@
     <PackageReference Include="SixLabors.ImageSharp" />
     <PackageReference Include="SSH.NET" PrivateAssets="compile" />
     <PackageReference Include="FFmpeg.AutoGen" Condition="'$(TargetFramework)' == '$(StrideFramework)'" />
+    <PackageReference Include="Microsoft.TemplateEngine.IDE" />
+    <PackageReference Include="Microsoft.TemplateEngine.Edge" />
+    <PackageReference Include="Microsoft.TemplateEngine.Abstractions" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\assets\Stride.Core.Assets\Stride.Core.Assets.csproj" />
